block metrics moved to common place, l1.rs removed, refactored l1.rs with alloy rs
