+++ crates/client/eth/Cargo.toml
@@ -16,7 +16,6 @@ targets = ["x86_64-unknown-linux-gnu"]
 # Deoxys
 dc-db = { workspace = true }
 dc-metrics = { workspace = true }
-dc-sync = { workspace = true }
 dp-convert = { workspace = true }
 dp-transactions = { workspace = true }
 dp-utils = { workspace = true }
