+++ crates/client/metrics/src/block_metrics/block_metrics.rs
@@ -1,6 +1,6 @@
 // use dc_metrics::{Gauge, MetricsRegistry, PrometheusError, F64};
 
-use crate::{Gauge, MetricsRegistry, PrometheusError, F64};
+use crate::{MetricsRegistry, PrometheusError, Gauge, F64};
 
 #[derive(Clone, Debug)]
 pub struct BlockMetrics {
@@ -13,7 +13,7 @@ pub struct BlockMetrics {
     pub transaction_count: Gauge<F64>,
     pub event_count: Gauge<F64>,
     // L1 network metrics
-    // gas price is also define in eth/client.rs but this would be the gas used in the block and it's price
+    pub l1_block_number: Gauge<F64>,
     pub l1_gas_price_wei: Gauge<F64>,
     pub l1_gas_price_strk: Gauge<F64>,
 }
@@ -30,6 +30,8 @@ impl BlockMetrics {
                 .register(Gauge::new("deoxys_l2_latest_sync_time", "Gauge for deoxys L2 latest sync time")?)?,
             l2_state_size: registry
                 .register(Gauge::new("deoxys_l2_state_size", "Gauge for node storage usage in GB")?)?,
+            l1_block_number: registry
+                .register(Gauge::new("deoxys_l1_block_number", "Gauge for deoxys L1 block number")?)?,
             transaction_count: registry
                 .register(Gauge::new("deoxys_transaction_count", "Gauge for deoxys transaction count")?)?,
             event_count: registry.register(Gauge::new("deoxys_event_count", "Gauge for deoxys event count")?)?,
