--- crates/client/sync/src/metrics/mod.rs
@@ -1,1 +1,0 @@
-pub mod block_metrics;
