+++ crates/node/src/service/mod.rs
@@ -1,4 +1,3 @@
-mod l1;
 pub mod rpc;
 pub mod sync;
